backend/requirements.txt
Flask==3.0.2
pymongo==4.6.3
python-dotenv==1.0.1
flask-cors==4.0.0